refactor: reduce boilerplate by moving controller/auditing test setup to ArmadilloControllerTest
